+++ AnorocMobileApp/AnorocMobileApp/AnorocMobileApp.csproj
@@ -195,6 +195,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <Folder Include="ViewModels\Dashboard" />
     <Folder Include="Views\Dashboard" />
   </ItemGroup>
 </Project>
